Moved VariableHolder out of VariableCodec for a later usage by immediate values.
